requirements.txt
# P2P Network Dependencies
# ========================

# Async HTTP client/server
aiohttp

# Async console input
aioconsole

# HTML parsing for web_reader agent
beautifulsoup4

# Environment configuration (.env files)
python-dotenv

# Cryptography (Ed25519, Curve25519, NaCl)
pynacl

# Colored terminal output
colorama

# Async SQLite for ledger
aiosqlite

# Async file operations
aiofiles

# Sandbox for contracts (optional)
RestrictedPython

# Distributed AI Inference
numpy>=1.24.0
chromadb
sentence-transformers
pypdf
transformers
torch
pillow
opencv-python
einops
timm
imagehash
insightface
onnxruntime-gpu
scikit-learn
presidio-analyzer
presidio-anonymizer
geoip2
spacy

# PyTorch Model Loading (optional, for real inference)
# torch>=2.0.0
# transformers>=4.35.0
# accelerate>=0.24.0
# safetensors>=0.4.0

# Web UI
nicegui>=1.4.0

# [PRODUCTION] System monitoring
psutil>=5.9.0

# Testing
pytest
pytest-asyncio
